backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary
gunicorn
django-cors-headers
djangorestframework